Extract Docker internals into its own module
